python-agent/pyproject.toml
[project]
name = "collider-agent"
version = "0.1.0"
description = "AI-powered agent for converting LaTeX Lagrangians to FeynRules model files"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Google ADK Framework
    "google-adk",
    # LLM Integration
    "litellm>=1.0.0",
    # Symbolic Math & Parsing
    "sympy>=1.12",
    "pyparsing>=3.1.0",
    # Web Scraping (for downloader utility)
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    # Data Handling
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    # Utilities
    "tqdm>=4.66.0",
    # Type hints and validation
    "typing-extensions>=4.8.0",
    # Mathematica Integration (for physics validation)
    "wolframclient>=1.1.7",
    # MadGraph Integration (for event generation)
    "pexpect>=4.8.0",
    # Cryptography (authlib dependency - use version compatible with newer cryptography)
    "authlib>=1.4.0",
    "magnus-sdk>=0.5.4",
    "pytest>=9.0.2",
]

[project.optional-dependencies]
# Optional: Semantic search for reference model matching
semantic = [
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
]

# Optional: LHE file parsing (we provide custom parser)
lhe = [
    "pylhe>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools", "utils"]

[dependency-groups]
dev = []

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--import-mode=importlib"

[tool.uv]


[tool.uv.sources]
# Lock file uses Tsinghua mirror for faster downloads in China
# This is configured in the uv.lock file automatically
